backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai>=1.6.1
google-generativeai==0.3.1
faiss-cpu==1.7.4
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
sentence-transformers==2.2.2

